requirements.txt
d20==1.1.2
discord==2.3.2
openai==0.28.0
python-dotenv==1.0.0
